--- a/internal/errors/std.go
+++ b/internal/errors/std.go
@@ -6,3 +6,8 @@ import "errors"
 func Is(err, target error) bool {
 	return errors.Is(err, target)
 }
+
+// As is a wrapper around the errors.As function.
+func As(err error, target interface{}) bool {
+	return errors.As(err, target)
+}