The `values.Value` interface represents a flux value and it is used by a
large number of runtime functions. The default implementation of the
basic values implements the data access functions, such as `Float()` and
`Int()`, for those respective types.

These access functions were deemed too complex to inline and so
performance was worse for accessing this data.

    values/values.go:69:6: cannot inline value.Float: function too complex: cost 135 exceeds budget 80

One reason is because `semantic.MonoType` required accessing the
flatbuffers data to determine the type nature. Instead of looking up the
type nature through a large function, this lookup has been moved to the
creation of the `semantic.MonoType` so that checking the nature is
accessing a member variable. For basic types, which generally use the
cached `semantic.BasicString`, this means that the type lookup is
incurred once instead of each time the value is accessed.

The other reason is because the panic function, `UnexpectedKind`, was
deemed too hairy by the go inliner. In order to optimize this function,
a new `unexpectedKind` error type was created and used to construct the
error. This `unexpectedKind` error type implements the `errors.As()`
function and can be accessed as a `*flux.Error`.

    err := values.UnexpectedKind(...)
    var ferr *flux.Error
    if errors.As(err, &ferr) {
        // As a flux error.
    }

This is the standard method of casting error types in go since 1.13.

Now, the `Float()` method is simpler and can be inlined.

    values/values.go:69:6: can inline value.Float with cost 31 as: method(value) func() float64 { CheckKind(v.t.Nature(), semantic.Float); return v.v.(float64) }

While the function itself and faster and can be inlined, there is still
a possibility that it will be faster but not inlined. This is because
`values.Value` is an interface and inlining isn't generally possible.

### Done checklist
- [ ] docs/SPEC.md updated
- [ ] Test cases written