@jsternberg 2021-01-14
Benchmark comparison:

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkValue_Str-12       16.8          4.00          -76.19%
BenchmarkValue_Bytes-12     17.4          4.38          -74.83%
BenchmarkValue_Int-12       15.3          3.36          -78.04%
BenchmarkValue_UInt-12      15.5          3.39          -78.13%
BenchmarkValue_Float-12     15.5          3.40          -78.06%
BenchmarkValue_Bool-12      13.4          3.43          -74.40%

benchmark                   old allocs     new allocs     delta
BenchmarkValue_Str-12       0              0              +0.00%
BenchmarkValue_Bytes-12     0              0              +0.00%
BenchmarkValue_Int-12       0              0              +0.00%
BenchmarkValue_UInt-12      0              0              +0.00%
BenchmarkValue_Float-12     0              0              +0.00%
BenchmarkValue_Bool-12      0              0              +0.00%

benchmark                   old bytes     new bytes     delta
BenchmarkValue_Str-12       0             0             +0.00%
BenchmarkValue_Bytes-12     0             0             +0.00%
BenchmarkValue_Int-12       0             0             +0.00%
BenchmarkValue_UInt-12      0             0             +0.00%
BenchmarkValue_Float-12     0             0             +0.00%
BenchmarkValue_Bool-12      0             0             +0.00%
```